requirements.txt
aniso8601==7.0.0
Django==2.2.13
django-filter==2.3.0
django-model-utils==4.0.0
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
python-dotenv==0.13.0
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
